<-- debugg migrate table

+++ database/migrations/2021_02_25_141618_create_sponsors.php
@@ -17,7 +17,7 @@ class CreateSponsors extends Migration
             $table->id();
             $table->unsignedBigInteger('apartment_id');
             $table->foreign('apartment_id')->references('id')->on('apartments')->onDelete('cascade');
-            $table->integer('type');
+            $table->unsignedBigInteger('type');
             $table->foreign('type')->references('id')->on('sponsors_type')->onDelete('cascade');
             $table->date('date_start');
             $table->date('date_end');
